refactor: extract `getOptionsForFile`
